requirements.txt
future==0.18.2
iso8601==1.0.2
numpy==1.21.4
opencv-python==4.5.4.60
pyserial==3.5
PyYAML==6.0
